feat(otel-core): extract CAS + trace-context + hex primitives from otel-scrape
Create packages/@overeng/otel-core (Rust lib) and move the registry-agnostic
primitives out of otel-scrape: content_address (CAS), context (TraceContext +
traceparent mint/join), and hex/hash helpers. otel-scrape re-imports them
one-way (pub use otel_core::context::TraceContext preserves the public API),
discharging decision 0009's "second Rust consumer" trigger.

Deferred to later PRs (saturated with the otel_scrape registry / entangled):
the OTLP exporter, span model, trust-gate (→ weaver-native fold with #882),
the wrap primitive (RunConfig/orchestration coupling), and build-id.

Wiring mirrors the otelite pattern (lib, no bin). Nix hermeticity: widen
otel-scrape's fileset root to the @overeng parent so the ../otel-core path dep
enters the sandbox. Verified by `nix build .#otel-core .#otel-scrape`.

97 tests preserved (94 otel-scrape + 3 otel-core); cargo:check / genie:check /
check:all green.
