+++ packages/@overeng/otel-core/.gitignore
@@ -1,0 +1,2 @@
+# Rust build artifacts
+/target/
