+++ packages/@overeng/otel-core/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "otel-core"
+version = "0.0.0"
+edition = "2021"
+description = "Registry-agnostic OpenTelemetry primitives: content-address CAS, W3C trace-context, hex/stable-hash"
+license = "MIT"
+
+[lib]
+name = "otel_core"
+path = "src/lib.rs"
+
+[dependencies]
+getrandom = "0.4"
+libc = "=0.2.186"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.11"
+
+[dev-dependencies]
+tempfile = "3"
+
+[profile.release]
+opt-level = 3
+lto = "thin"
+strip = "symbols"
